# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ATLAS BI - Commercial Dashboard API for NIADCON
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Consultant {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  region?: string | null;
  active: boolean;
  createdAt: string;
}

export interface ConsultantInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  region?: string;
  active?: boolean;
}

export interface ConsultantUpdate {
  name?: string;
  email?: string;
  phone?: string;
  region?: string;
  active?: boolean;
}

export interface Sale {
  id: number;
  consultantId: number;
  /** @nullable */
  consultantName?: string | null;
  product: string;
  amount: number;
  quantity: number;
  saleDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SaleInput {
  consultantId: number;
  /** @minLength 1 */
  product: string;
  amount: number;
  quantity: number;
  saleDate: string;
  notes?: string;
}

export interface SaleUpdate {
  consultantId?: number;
  product?: string;
  amount?: number;
  quantity?: number;
  saleDate?: string;
  notes?: string;
}

export interface Goal {
  id: number;
  /** @nullable */
  consultantId?: number | null;
  /** @nullable */
  consultantName?: string | null;
  month: number;
  year: number;
  targetAmount: number;
  /** @nullable */
  targetQuantity?: number | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface GoalInput {
  consultantId?: number;
  month: number;
  year: number;
  targetAmount: number;
  targetQuantity?: number;
  description?: string;
}

export interface GoalUpdate {
  targetAmount?: number;
  targetQuantity?: number;
  description?: string;
}

export interface DashboardSummary {
  totalSales: number;
  totalQuantity: number;
  totalConsultants: number;
  activeConsultants: number;
  goalAmount: number;
  goalAchievementPercent: number;
  /** @nullable */
  topConsultantName?: string | null;
  /** @nullable */
  topConsultantAmount?: number | null;
  /** @nullable */
  previousMonthSales?: number | null;
  /** @nullable */
  growthPercent?: number | null;
}

export interface RankingEntry {
  position: number;
  consultantId: number;
  consultantName: string;
  totalAmount: number;
  totalQuantity: number;
  /** @nullable */
  goalAmount?: number | null;
  /** @nullable */
  goalAchievementPercent?: number | null;
}

export interface MonthlyProduction {
  month: number;
  monthName: string;
  totalAmount: number;
  totalQuantity: number;
  /** @nullable */
  goalAmount?: number | null;
}

export type ListSalesParams = {
/**
 * @nullable
 */
consultantId?: number | null;
/**
 * Month number 1-12
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
};

export type ListGoalsParams = {
/**
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
};

export type GetDashboardSummaryParams = {
/**
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
};

export type GetDashboardRankingParams = {
/**
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
/**
 * Top N consultants to return
 * @nullable
 */
limit?: number | null;
};

export type GetProductionChartParams = {
/**
 * @nullable
 */
year?: number | null;
};