# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ATLAS BI - Commercial Dashboard API for NIADCON
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consultants:
    get:
      operationId: listConsultants
      tags: [consultants]
      summary: List all consultants
      responses:
        "200":
          description: List of consultants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultant"
    post:
      operationId: createConsultant
      tags: [consultants]
      summary: Create a new consultant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultantInput"
      responses:
        "201":
          description: Consultant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
  /consultants/{id}:
    get:
      operationId: getConsultant
      tags: [consultants]
      summary: Get a consultant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultant found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateConsultant
      tags: [consultants]
      summary: Update a consultant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
    delete:
      operationId: deleteConsultant
      tags: [consultants]
      summary: Delete a consultant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary KPIs
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/ranking:
    get:
      operationId: getDashboardRanking
      tags: [dashboard]
      summary: Get sales ranking by consultant
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
          description: "Top N consultants to return"
      responses:
        "200":
          description: Ranked consultants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingEntry"
  /dashboard/production-chart:
    get:
      operationId: getProductionChart
      tags: [dashboard]
      summary: Get monthly production data for charts
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Production chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyProduction"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - id
        - name
        - email
        - role
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    Consultant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - active
        - createdAt
    ConsultantInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        region:
          type: string
        active:
          type: boolean
      required:
        - name
        - email
    ConsultantUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        region:
          type: string
        active:
          type: boolean
...
    DashboardSummary:
      type: object
      properties:
        totalSales:
          type: number
        totalQuantity:
          type: integer
        totalConsultants:
          type: integer
        activeConsultants:
          type: integer
        goalAmount:
          type: number
        goalAchievementPercent:
          type: number
        topConsultantName:
          type: ["string", "null"]
        topConsultantAmount:
          type: ["number", "null"]
        previousMonthSales:
          type: ["number", "null"]
        growthPercent:
          type: ["number", "null"]
      required:
        - totalSales
        - totalQuantity
        - totalConsultants
        - activeConsultants
        - goalAmount
        - goalAchievementPercent
    RankingEntry:
      type: object
      properties:
        position:
          type: integer
        consultantId:
          type: integer
        consultantName:
          type: string
        totalAmount:
          type: number
        totalQuantity:
          type: integer
        goalAmount:
          type: ["number", "null"]
        goalAchievementPercent:
          type: ["number", "null"]
      required:
        - position
        - consultantId
        - consultantName
        - totalAmount
        - totalQuantity
    MonthlyProduction:
      type: object
      properties:
        month:
          type: integer
        monthName:
          type: string
        totalAmount:
          type: number
        totalQuantity:
          type: integer
        goalAmount:
          type: ["number", "null"]
      required:
        - month
        - monthName
        - totalAmount
        - totalQuantity