# lib/api-zod/src/generated/types/consultantUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ATLAS BI - Commercial Dashboard API for NIADCON
 * OpenAPI spec version: 0.1.0
 */

export interface ConsultantUpdate {
  name?: string;
  email?: string;
  phone?: string;
  region?: string;
  active?: boolean;
}