# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ATLAS BI - Commercial Dashboard API for NIADCON
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalSales: number;
  totalQuantity: number;
  totalConsultants: number;
  activeConsultants: number;
  goalAmount: number;
  goalAchievementPercent: number;
  /** @nullable */
  topConsultantName?: string | null;
  /** @nullable */
  topConsultantAmount?: number | null;
  /** @nullable */
  previousMonthSales?: number | null;
  /** @nullable */
  growthPercent?: number | null;
}